requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
colorama==0.4.4
fastapi==0.75.0
Flask==2.0.3
gunicorn==20.1.0
h11==0.13.0
idna==3.3
imageio==2.16.1
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.1.1
networkx==2.7.1
numpy==1.22.3
opencv-python-headless==4.5.5.64
packaging==21.3
Pillow==9.0.1
pydantic==1.9.0
pyparsing==3.0.7
python-multipart==0.0.5
PyWavelets==1.3.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
threadpoolctl==3.1.0
tifffile==2022.3.16
typing_extensions==4.1.1
uvicorn==0.17.6
Werkzeug==2.0.3
